requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
colorlog==6.8.0
Django==4.2.9
django-filter==23.5
django-ninja==1.1.0
djangorestframework==3.14.0
importlib-metadata==7.0.1
JSON-log-formatter==0.5.2
Markdown==3.5.2
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-core==2.14.6
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.9.0
zipp==3.17.0
